Move logout api call to calls
